Revert "feat: triage→Conversion queue split for fast-path concurrency (#282)"
This reverts commit 46325fa99db07ea65f2fc141cc088abae616bb21.
