+++ asset-bundle-converter/Packages/manifest.json
@@ -2,7 +2,7 @@
   "dependencies": {
     "com.atteneder.draco": "4.0.2",
     "com.atteneder.ktx": "2.2.3",
-    "com.cysharp.unitask": "https://github.com/Cysharp/UniTask.git?path=src/UniTask/Assets/Plugins/UniTask#305695ad5d194dda0d6b00777c2459202a1ef4ea",
+    "com.cysharp.unitask": "https://github.com/Cysharp/UniTask.git?path=src/UniTask/Assets/Plugins/UniTask",
     "com.unity.assetbundlebrowser": "https://github.com/Unity-Technologies/AssetBundles-Browser.git",
     "com.unity.ide.rider": "3.0.38",
     "com.unity.ide.visualstudio": "2.0.23",
