--- consumer-server/src/adapters/filesystem/index.ts
@@ -1,2 +1,0 @@
-export { createFilesystemComponent } from './component'
-export type { IFilesystemComponent, FilesystemComponentOptions } from './types'
