--- consumer-server/src/adapters/unity-runner/index.ts
@@ -1,2 +1,0 @@
-export { createUnityRunnerComponent } from './component'
-export type { IUnityRunnerComponent, RunConversionOptions, RunLodsConversionOptions } from './types'
