--- consumer-server/src/logic/conversion-orchestrator/index.ts
@@ -1,3 +1,0 @@
-export { createConversionOrchestratorComponent } from './component'
-export type { IConversionOrchestratorComponent, Platform } from './types'
-export { ConversionQueueRepublishFailedError } from './errors'
