--- consumer-server/src/logic/scenes/index.ts
@@ -1,2 +1,0 @@
-export { createScenesComponent } from './component'
-export type { IScenesComponent, Manifest, ProbeArgs, ProbeOutcome, UploadFastPathArgs } from './types'
