+++ consumer-server/src/service.ts
@@ -1,6 +1,10 @@
 import { Lifecycle } from '@well-known-components/interfaces'
 import { setupRouter } from './controllers/routes'
+import { executeConversion, executeLODConversion } from './logic/conversion-task'
+import checkDiskSpace from 'check-disk-space'
 import { AppComponents, GlobalContext, TestComponents } from './types'
+import { AssetBundleConversionFinishedEvent, Events } from '@dcl/schemas'
+import { getAbVersionEnvName } from './utils'
 
 // this function wires the business logic (adapters & controllers) with the components (ports)
 export async function main(program: Lifecycle.EntryPointParameters<AppComponents | TestComponents>) {
@@ -21,42 +25,96 @@ export async function main(program: Lifecycle.EntryPointParameters<AppComponents
   // start ports: db, listeners, synchronizations, etc
   await startComponents()
 
-  const conversionLogger = components.logs.getLogger('conversion-loop')
+  const logger = components.logs.getLogger('main-loop')
 
-  // Triage loop. When `FAST_PATH_TRIAGE_ENABLED` is off (default), the
-  // orchestrator's processIncomingJob runs full executeConversion inline
-  // (today's behavior). When on, it runs executeTriagePass and either
-  // fast-paths or republishes to the Conversion queue. See
-  // `logic/conversion-orchestrator/component.ts` for the full decision tree.
-  //
-  // No disk-pressure gate here: in fast-path mode the triage loop only does
-  // probe-and-republish work (no Unity spawn, no scene materialisation), so
-  // it stays useful when disk is tight. The conversion loop owns the
-  // disk-pressure shutdown — that's where the actual writes happen, and the
-  // inline-fallback path runs `executeConversion` via the orchestrator, so
-  // disk-pressure-triggered shutdown still fires on this pod when needed.
   components.runner.runTask(async (opt) => {
-    while (opt.isRunning) {
-      await components.triageTaskQueue.consumeAndProcessJob(async (job, _message, opts) => {
-        await components.conversionOrchestrator.processIncomingJob(job, opts.isPriority)
-      })
-    }
-  })
+    const platform = (await components.config.requireString('PLATFORM')).toLocaleLowerCase() as
+      | 'windows'
+      | 'mac'
+      | 'webgl'
+    const $BUILD_TARGET = await components.config.requireString('BUILD_TARGET')
+    const abVersionEnvName = getAbVersionEnvName($BUILD_TARGET)
+    const $AB_VERSION = await components.config.requireString(abVersionEnvName)
 
-  // Conversion loop. Always runs; drains the Conversion queue regardless of
-  // the kill switch. When the switch is off, no new messages land here, so
-  // this loop sits idle. On revert, any residual messages drain naturally.
-  components.runner.runTask(async (opt) => {
     while (opt.isRunning) {
-      if (await components.filesystem.isBelowMinimum()) {
-        conversionLogger.warn('Stopping program due to lack of disk space (conversion loop)')
+      if (await machineRanOutOfSpace(components)) {
+        logger.warn('Stopping program due to lack of disk space')
         void program.stop()
         return
       }
 
-      await components.conversionTaskQueue.consumeAndProcessJob(async (job, _message, _opts) => {
-        await components.conversionOrchestrator.processConversionJob(job)
+      await components.taskQueue.consumeAndProcessJob(async (job, _message) => {
+        // Skip non-conversion payloads that may leak into the queue (e.g. world_undeployment).
+        if (!job?.entity?.entityId) {
+          logger.warn('Skipping job with no entity.entityId — not a conversion job', {
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            type: (job as any)?.type,
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            subType: (job as any)?.subType
+          })
+          return
+        }
+
+        let statusCode: number
+        try {
+          components.metrics.increment('ab_converter_running_conversion')
+
+          // Increment version if doISS is true
+          let versionToUse = $AB_VERSION
+          if (job.doISS) {
+            versionToUse = 'v2004'
+          }
+
+          if (job.lods) {
+            statusCode = await executeLODConversion(components, job.entity.entityId, job.lods, versionToUse)
+          } else {
+            statusCode = await executeConversion(
+              components,
+              job.entity.entityId,
+              job.contentServerUrls![0],
+              job.force,
+              job.animation,
+              job.doISS,
+              versionToUse
+            )
+          }
+
+          const eventToPublish: AssetBundleConversionFinishedEvent = {
+            type: Events.Type.ASSET_BUNDLE,
+            subType: Events.SubType.AssetBundle.CONVERTED,
+            key: `${job.entity.entityId}-${platform}`,
+            timestamp: Date.now(),
+            metadata: {
+              platform: platform,
+              entityId: job.entity.entityId,
+              isLods: !!job.lods,
+              isWorld:
+                !!job.contentServerUrls &&
+                job.contentServerUrls.length > 1 &&
+                job.contentServerUrls[0].includes('worlds-content-server'),
+              statusCode,
+              version: versionToUse
+            }
+          }
+
+          await components.publisher.publishMessage(eventToPublish)
+        } finally {
+          components.metrics.decrement('ab_converter_running_conversion')
+        }
       })
     }
   })
+}
+
+async function machineRanOutOfSpace(components: Pick<AppComponents, 'metrics'>) {
+  const diskUsage = await checkDiskSpace('/')
+  const free = diskUsage.free
+
+  components.metrics.observe('ab_converter_free_disk_space', {}, free)
+
+  if (free / 1e9 < 1 /* less than 1gb */) {
+    return true
+  }
+
+  return false
 }
