+++ consumer-server/test/e2e/conversion.spec.ts
@@ -48,12 +48,6 @@ process.env.PLATFORM = $BUILD_TARGET
 process.env.ASSET_REUSE_ENABLED = 'true'
 // AWS_SNS_ARN is required by the publisher component at init time
 process.env.AWS_SNS_ARN = process.env.AWS_SNS_ARN || 'arn:aws:sns:us-east-1:000000000000:e2e-test'
-// AWS_REGION is required by the SNS v3 client to sign the publish call. Without
-// it, every post-conversion publish fails with "Region is missing" before any
-// network attempt — noisy in CI logs and obscures a real SNS regression. The
-// fake ARN above means the request never reaches a real SNS endpoint anyway,
-// so any region that lets the SDK construct the signed request will do.
-process.env.AWS_REGION = process.env.AWS_REGION || 'us-east-1'
 
 const MOCK_S3_BASE = '/tmp/e2e-mock-s3'
 const BUCKET_NAME = 'CDN_BUCKET' // default when CDN_BUCKET env is unset
@@ -188,7 +182,7 @@ test('when converting scenes end-to-end via queue-task', ({ components }) => {
     // Publish directly to the in-memory queue — bypasses the HTTP handler
     // (which requires SNS) but still exercises the queue consumer loop in
     // service.ts that calls executeConversion.
-    await components.triageTaskQueue.publish({
+    await components.taskQueue.publish({
       entity: {
         entityId,
         authChain: [{ type: AuthLinkType.SIGNER, payload: '0x0000000000000000000000000000000000000000', signature: '' }]
